PMO: Don't block pmo for large types - rather only block expansion of tuples
